fix(ebpf): drop unnecessary uint64() conversion on stat.Dev
unconvert lint: stat.Dev is already uint64 on Linux's syscall.Stat_t
(both amd64 and arm64), so the explicit conversion is redundant.

+++ pkg/ebpf/uprobe.go
@@ -1165,7 +1165,7 @@ func walkTLSLibrariesUnder(rootPrefix string) []string {
 			if !ok {
 				return nil
 			}
-			id := fileID{dev: uint64(stat.Dev), ino: stat.Ino}
+			id := fileID{dev: stat.Dev, ino: stat.Ino}
 			if seen[id] {
 				return nil
 			}
